fix(api): validate block number range in parseBlockNumber
parseBlockNumber previously returned BigInteger, letting callers narrow
with .longValue() and silently wrap past Long.MAX_VALUE. Return long
directly and enforce two invariants at the API boundary so corrupted
inputs never reach downstream block lookups:

- signum check rejects negative inputs (both "-1" and "0x-1"),
  a protocol-level constraint block numbers can never violate
- longValueExact rejects values that overflow signed 64-bit,
  including uint64 wraparounds like 0xffffffffffffffff

Callers in Wallet and TronJsonRpcImpl drop the obsolete .longValue()
call; the other three call sites already discarded the return value.

Regression tests cover negatives, 0x7fffffffffffffff (max long),
0x8000000000000000 (just past), and 0xffffffffffffffff (uint64).

+++ framework/src/main/java/org/tron/core/Wallet.java
@@ -743,7 +743,7 @@ public class Wallet {
     } else if (PENDING_STR.equalsIgnoreCase(id)) {
       throw new JsonRpcInvalidParamsException(TAG_PENDING_SUPPORT_ERROR);
     } else {
-      long blockNumber = parseBlockNumber(id).longValue();
+      long blockNumber = parseBlockNumber(id);
       return getBlockByNum(blockNumber);
     }
   }
