+++ framework/src/main/java/org/tron/core/services/jsonrpc/TronJsonRpcImpl.java
@@ -957,7 +957,7 @@ public class TronJsonRpcImpl implements TronJsonRpc, Closeable {
           throw new JsonRpcInvalidParamsException(JSON_ERROR);
         }
 
-        long blockNumber = parseBlockNumber(blockNumOrTag).longValue();
+        long blockNumber = parseBlockNumber(blockNumOrTag);
 
         if (wallet.getBlockByNum(blockNumber) == null) {
           throw new JsonRpcInternalException(NO_BLOCK_HEADER);
